dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
click==7.0                # via pip-tools
coverage==4.5.4           # via pytest-cov
importlib-metadata==0.23  # via pluggy, pytest
more-itertools==7.2.0     # via pytest, zipp
packaging==19.2           # via pytest
pip-tools==4.2.0
pluggy==0.13.0            # via pytest
py==1.8.0                 # via pytest
pyparsing==2.4.2          # via packaging
pytest-cov==2.7.1
pytest-django==3.6.0
pytest==5.2.1
six==1.12.0               # via packaging, pip-tools
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata
